requirements.txt
# Core dependencies
requests==2.31.0
python-dotenv==1.0.0
gspread==5.12.0
oauth2client==4.1.3

# AI/ML dependencies
openai==1.3.0
anthropic==0.7.0

# Pinterest API
pinterest-api==1.0.0

# Shopify API
shopify-python-api==12.0.0

# Data processing
pandas==2.1.0
numpy==1.24.0

# Web framework
flask==3.0.0
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.0

# Task queue
celery==5.3.0
redis==5.0.0

# Monitoring
prometheus-client==0.19.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3